# src/doc/swagger_doc.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Ammo Varejo RestAPI",
    "description": "Teste prático para o grupo Ammo Varejo."
  },
  "host": "ammo-varejo-api.onrender.com",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/product/searchProduct": {
      "get": {
        "tags": [
          "product"
        ],
        "description": "Pesquisar produtos",
        "parameters": [
          {
            "name": "term",
            "description": "Termo para buscar o produto (Exemplo: Kit Cama)",
            "type": "string",
            "in": "query"
          },
          {
            "name": "page",
            "description": "Página dos itens",
            "required": true,
            "type": "integer",
            "in": "query"
          },
          {
            "name": "itens",
            "description": "Quantidade de itens a buscar",
            "required": true,
            "type": "integer",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idProduct": {
                        "type": "string",
                        "example": "0"
                      },
                      "name": {
                        "type": "string",
                        "example": "nome"
                      },
                      "description": {
                        "type": "string",
                        "example": "descricao"
                      },
                      "promotionalPrice": {
                        "type": "string",
                        "example": "preco promocional"
                      },
                      "originalPrice": {
                        "type": "string",
                        "example": "preco original"
                      },
                      "idCategory": {
                        "type": "string",
                        "example": "id da categoria"
                      },
                      "images": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "idImage": {
                              "type": "number",
                              "example": 0
                            },
                            "idProduct": {
                              "type": "number",
                              "example": 0
                            },
                            "src": {
                              "type": "string",
                              "example": "caminho para a imagem"
                            }
                          }
                        }
                      },
                      "category": {
                        "type": "object",
                        "properties": {
                          "idCategory": {
                            "type": "string",
                            "example": "id categoria"
                          },
                          "name": {
                            "type": "string",
                            "example": "nome da categoria"
                          }
                        }
                      }
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Sucesso"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Erro"
          }
        }
      }
    },
    "/category/searchByTerm": {
      "get": {
        "tags": [
          "category"
        ],
        "description": "Pesquisar categorias",
        "parameters": [
          {
            "name": "term",
            "description": "Termo para buscar a categoria (Exemplo: Cama)",
            "type": "string",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idCategory": {
                        "type": "number",
                        "example": 0
                      },
                      "name": {
                        "type": "string",
                        "example": "nome"
                      }
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Sucesso"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Erro"
          }
        }
      }
    }
  }
}